% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cached_data.R
\name{get_cached_data}
\alias{get_cached_data}
\title{get_dataset}
\usage{
get_cached_data(username, api_key, file_to_download, target = NULL)
}
\arguments{
\item{username}{Your dataservice user name}

\item{api_key}{Your API key}

\item{file_to_download}{The name of the file to retrieve.}

\item{target}{Path to the location, at which to store the file. If NULL, the file will be
saved to the working directory.}
}
\description{
Download a predefined set of time series.
}
\examples{
httr_response <- get_cached_data(userame, apikey, "some_file.xlsx")
get_cached_data(username, apikey, "some_file.xlsx", target="C:/data/some_other_name.xlsx")
}
