% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{Download Pre-Defined Dataset}
\usage{
get_dataset(set_name, api_key = NULL, show_progress = FALSE)
}
\arguments{
\item{set_name}{The name of the set you wish to download. For a list of available sets, go to ["list_available_sets"]}

\item{api_key}{\strong{character} Your API key. An API key is needed if accessing non-public timeseries. 
If downloading public timeseries parameter can be omitted.}

\item{show_progress}{If set to true, shows a progress bar of the data being downloaded.}
}
\description{
This function is deprecated. Use \code{\link{get_collection}} instead.
Download a predefined set of time series.
}
\examples{
get_dataset("ds_kmi_mixed_freq",show_progress = TRUE)
}
