% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{get_metadata}
\usage{
get_metadata(ts_keys, locale = c("en", "de", "fr", "it"))
}
\arguments{
\item{ts_keys}{A vector of time series keys.}

\item{locale}{The language in which to return the metadata (if available)}
}
\value{
A named list of lists containing metadata.
}
\description{
Download metadata relating to the specified keys.
}
