% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_to_ts.R
\name{json_to_ts}
\alias{json_to_ts}
\title{Helper to turn api-returned json (lists) into lists of ts objects}
\usage{
json_to_ts(json_data)
}
\description{
Helper to turn api-returned json (lists) into lists of ts objects
}
