% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_available_collections.R
\name{list_available_collections}
\alias{list_available_collections}
\title{List Available collection Names and Their Description}
\usage{
list_available_collections(api_key = NULL)
}
\arguments{
\item{api_key}{\strong{character} Your API key. An API key is needed if accessing non-public timeseries. 
If downloading public timeseries parameter can be omitted.}
}
\value{
A data frame with the names, descriptions and public availability of collections. These collections can be downloaded via get_collection.
}
\description{
collections are pre-defined lists of time series. collections are the convenient alternative to concatenating many series in an URL.
}
\examples{
list_available_collections()
}
