% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_available_sets.R
\name{list_available_sets}
\alias{list_available_sets}
\title{List Available Set Names and Their Description}
\usage{
list_available_sets(use_tempfile = FALSE)
}
\arguments{
\item{use_tempfile}{Store downloaded data to a temporary file and read from there.
This option can be used to work around security measures preventing reading directly into RStudio's memory.}
}
\value{
A data frame with the names, descriptions and public availability of datasets. These sets can be downloaded via get_dataset.
}
\description{
Sets are pre-defined lists of time series. Sets are the convenient alternative to concatenating many series in an URL.
}
\examples{
list_available_sets()
}
