% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_cached_files.R
\name{list_cached_files}
\alias{list_cached_files}
\title{Show Available Files for Specific User}
\usage{
list_cached_files(username, api_key)
}
\arguments{
\item{username}{Your dataservice user name}

\item{api_key}{Your API key}
}
\description{
Show Available Files for Specific User
}
\examples{
# Not run
# available_files <- list_cached_files(username, apikey)
}
