% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_keys_in_set.R
\name{list_keys_in_set}
\alias{list_keys_in_set}
\title{List All Keys in a Set}
\usage{
list_keys_in_set(setname, use_tempfile = FALSE)
}
\arguments{
\item{setname}{The name of the set}

\item{use_tempfile}{Store downloaded data to a temporary file and read from there.
This option can be used to work around security measures preventing reading directly into RStudio's memory.}
}
\value{
If a single set name is provided, a vector of time series keys.
If multiple set names are provided, a list of vectors of time series keys.
}
\description{
List the keys of all time series in a set. To learn more about specific keys, use get_metadata.
}
\examples{
list_keys_in_set("ds_kmi_mixed_freq")
}
