% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_public_keys.R
\name{list_public_keys}
\alias{list_public_keys}
\title{List Keys for All Publicly Available Time Series}
\usage{
list_public_keys(use_tempfile = FALSE)
}
\arguments{
\item{use_tempfile}{Store downloaded data to a temporary file and read from there.
This option can be used to work around security measures preventing reading directly into RStudio's memory.}
}
\value{
An array of all public time series keys.
}
\description{
List Keys for All Publicly Available Time Series
}
\examples{
list_public_keys()
}
