% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_docs.R
\name{param_defs}
\alias{param_defs}
\title{Common parameters}
\arguments{
\item{api_key}{\strong{character} Your API key. An API key is needed if accessing non-public timeseries. 
If downloading public timeseries parameter can be omitted.}
}
\description{
Common parameters
}
