% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_legacy_key.R
\name{translate_legacy_key}
\alias{translate_legacy_key}
\title{Get a Current Time Series Identifier Given Legacy Keys}
\usage{
translate_legacy_key(legacy_keys, show_progress = T)
}
\arguments{
\item{legacy_keys}{A vector of KOF legacy keys}

\item{show_progress}{logical show progress bar bey shown? Defaults to TRUE.}
}
\value{
A named list of time series keys
}
\description{
Get new keys given old time series keys. Some legacy keys match more than one new key as more data is stored. This functions tries to return a single new key per old key. It always returns run 1, Noga08, mixed frequency keys.
}
\examples{
translate_legacy_key("chimt_d_f7_s01")
}
