% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_legacy_keys.R
\name{translate_legacy_keys}
\alias{translate_legacy_keys}
\title{translate_legacy_keys}
\usage{
translate_legacy_keys(legacy_keys, chunksize = 100)
}
\arguments{
\item{legacy_keys}{A vector of legacy keys for which to find new keys}

\item{chunksize}{integer to set chunksize in order to avoid URLs that are too long.}
}
\value{
A list of vectors of new ts_keys
}
\description{
Gives ts_keys for the given legacy keys.
}
