% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cached_file.R
\name{download_cached_file}
\alias{download_cached_file}
\title{get_dataset}
\usage{
download_cached_file(username, api_key, file_to_download, target = NULL)
}
\arguments{
\item{username}{Your dataservice user name}

\item{api_key}{\strong{character} Your API key. An API key is needed if accessing non-public timeseries. 
If downloading public timeseries parameter can be omitted.}

\item{file_to_download}{The name of the file to retrieve.}

\item{target}{Path to the location, at which to store the file. If NULL, the file will be
saved to the working directory.}
}
\description{
Download a predefined set of time series.
}
\examples{
f <-  download_cached_file("kofdatapkg", "313984fcd9f343d3961891319b0ed321",
"empty.txt",file.path(tempdir(),"empty.txt"))
}
