% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_collection.R
\name{get_collection}
\alias{get_collection}
\title{Download Pre-Defined collection}
\usage{
get_collection(collection_name, api_key = NULL, show_progress = FALSE)
}
\arguments{
\item{collection_name}{The name of the collection you wish to download. For a list of available collections, go to ["list_available_collections"]}

\item{api_key}{\strong{character} Your API key. An API key is needed if accessing non-public timeseries. 
If downloading public timeseries parameter can be omitted.}

\item{show_progress}{If collection to true, shows a progress bar of the data being downloaded.}
}
\description{
Download a predefined collection of time series.
}
\examples{
get_collection("ds_kmi_mixed_freq",show_progress = TRUE)
}
