% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{Download Pre-Defined Dataset}
\usage{
get_dataset(set_name, api_key = NULL, show_progress = FALSE)
}
\arguments{
\item{set_name}{The name of the set you wish to download. For a list of available sets, go to [TODO]}

\item{api_key}{Your API key. This is only needed if accessing non-public time series. Defaults to NULL (public data).}

\item{show_progress}{If set to true, shows a progress bar of the data being downloaded.}
}
\description{
Download a predefined set of time series.
}
\examples{
get_dataset("ds_kmi_mixed_freq",show_progress = TRUE)
}
