% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_legacy_key.R
\name{get_legacy_key}
\alias{get_legacy_key}
\title{Get a Former Time Series Identifier from KOF Database}
\usage{
get_legacy_key(ts_keys)
}
\arguments{
\item{ts_keys}{A vector of time series keys}
}
\value{
A named list of legacy keys
}
\description{
Get legacy keys given new time series keys. Note that not all records need to have 
a legacy key because there are more new keys than old keys.
}
