% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_series.R
\name{get_time_series}
\alias{get_time_series}
\title{Get Time Series form KOF Datenservice API}
\usage{
get_time_series(ts_keys, api_key = NULL, show_progress = FALSE)
}
\arguments{
\item{ts_keys}{A vector of timeseries keys}

\item{api_key}{Your API key. This is only needed if accessing non-public time series.}

\item{show_progress}{If set to true, shows a progress bar of the data being downloaded.
year-month-day, otherwise only year and month.}
}
\description{
Download time series data from the KOF web API.
To explore the available data and find the keys to series you are interested
in, run \code{\link{start_key_explorer}}.
}
\examples{
get_time_series("kofbarometer")
}
