% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_available_sets.R
\name{list_available_sets}
\alias{list_available_sets}
\title{list_available_sets}
\usage{
list_available_sets()
}
\value{
A data frame with the names, descriptions and public availability of datasets. These sets can be downloaded via get_dataset.
}
\description{
Get a list of all available data sets.
}
