% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_available_sets.R
\name{list_available_sets}
\alias{list_available_sets}
\title{List Available Set Names and Their Description}
\usage{
list_available_sets(api_key = NULL)
}
\arguments{
\item{api_key}{\strong{character} Your API key. An API key is needed if accessing non-public timeseries. 
If downloading public timeseries parameter can be omitted.}
}
\value{
A data frame with the names, descriptions and public availability of datasets. These sets can be downloaded via get_dataset.
}
\description{
This function is deprecated. Use \code{\link{list_available_collections}} instead.
Sets are pre-defined lists of time series. Sets are the convenient alternative to concatenating many series in an URL.
}
\examples{
list_available_sets()
}
