% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_cached_files.R
\name{list_cached_files}
\alias{list_cached_files}
\title{get_dataset}
\usage{
list_cached_files(username, api_key)
}
\arguments{
\item{username}{Your dataservice user name}

\item{api_key}{Your API key}
}
\description{
Download a predefined set of time series.
}
\examples{
available_files <- get_cached_data(username, apikey)
}
