% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_keys_in_collection.R
\name{list_keys_in_collection}
\alias{list_keys_in_collection}
\title{List All Keys in a Collection}
\usage{
list_keys_in_collection(collectionname, api_key = NULL)
}
\arguments{
\item{collectionname}{The name of the collection}

\item{api_key}{\strong{character} Your API key. An API key is needed if accessing non-public timeseries. 
If downloading public timeseries parameter can be omitted.}
}
\value{
If a single collection name is provided, a vector of time series keys.
If multiple collection names are provided, a list of vectors of time series keys.
}
\description{
List the keys of all time series in a collection. To learn more about specific keys, use get_metadata.
}
\examples{
list_keys_in_collection("ds_kmi_mixed_freq")
}
