% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_public_keys.R
\name{list_public_keys}
\alias{list_public_keys}
\title{List Keys for All Publicly Available Time Series}
\usage{
list_public_keys()
}
\value{
An array of all public time series keys.
}
\description{
This function will be removed from future versions of kofdata.
}
