% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_legacy_keys.R
\name{translate_legacy_keys}
\alias{translate_legacy_keys}
\title{translate_legacy_keys}
\usage{
translate_legacy_keys(legacy_keys, chunksize = 100)
}
\arguments{
\item{legacy_keys}{A vector of legacy keys for which to find new keys}

\item{chunksize}{integer to set chunksize in order to avoid URLs that are too long.}
}
\value{
A list of vectors of new ts_keys
}
\description{
Gives ts_keys for the given legacy keys.
Since there is no one to one mapping from old to new keys, it is possible that multiple new keys are returned for an old key.
Note: Not all possible new keys are returned. The keys returned by this function should serve as a starting point to find the
specific new keys that are needed. For the key explorer to find the exact keys, see \code{\link{start_key_explorer}}.
}
